Initial

+++ backend/controllers/products.controller.js
@@ -39,12 +39,16 @@ export const createProduct = async(req, res) =>{
 export const deleteProduct = async(req, res) =>{
     const { id } = req.params; 
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid product id" });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product deleted"});
     }
     catch (error){
-        res.status(404).json({ success: false, message: "Product not found"});
+        res.status(500).json({ success: false, message: "Product not in system"});
     }
 };
